chore: rename the API project to StudentAPI
Makes a clearer distinction between the API and the Web front end
