+++ api/StudentApi.Mediatr/InjectionPlaceholder.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace interview_test_angular.Mediatr
+namespace StudentApi.Mediatr
 {
     public class InjectionPlaceholder
     {
