+++ api/StudentApi.Mediatr/StudentApi.Mediatr.csproj
@@ -2,12 +2,17 @@
 
   <PropertyGroup>
     <TargetFramework>net5.0</TargetFramework>
-    <RootNamespace>interview_test_angular.Models</RootNamespace>
+    <RootNamespace>StudentApi.Mediatr</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="MediatR" Version="10.0.1" />
     <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="10.0.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\StudentApi.Models\StudentApi.Models.csproj" />
+    <ProjectReference Include="..\StudentApi.Services\StudentApi.Services.csproj" />
   </ItemGroup>
 
 </Project>
