+++ api/StudentApi.Services/StudentApi.Services.csproj
@@ -1,0 +1,12 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+    <RootNamespace>StudentApi.Services</RootNamespace>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\StudentApi.Models\StudentApi.Models.csproj" />
+  </ItemGroup>
+
+</Project>
