+++ api/StudentApi.Services/StudentsService.cs
@@ -1,11 +1,11 @@
-﻿using interview_test_angular.Models.Students;
+﻿using StudentApi.Models.Students;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace interview_test_angular.Services
+namespace StudentApi.Services
 {
     public class StudentsService : IStudentsService
     {
