+++ api/StudentApi/Pages/_ViewImports.cshtml
@@ -1,0 +1,3 @@
+﻿@using StudentApi
+@namespace StudentApi.Pages
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
