--- interview-test-angular/interview-test-angular.Services/interview-test-angular.Services.csproj
@@ -1,12 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
-    <RootNamespace>interview_test_angular.Services</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\interview-test-angular.Models\interview-test-angular.Models.csproj" />
-  </ItemGroup>
-
-</Project>
