--- interview-test-angular/interview-test-angular/Pages/_ViewImports.cshtml
@@ -1,3 +1,0 @@
-﻿@using interview_test_angular
-@namespace interview_test_angular.Pages
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
