+++ web/README.md
@@ -1,4 +1,4 @@
-# interview_test_angular
+# StudentApi
 
 This project was generated with [Angular CLI](https://github.com/angular/angular-cli) version 6.0.0.
