+++ web/angular.json
@@ -3,7 +3,7 @@
   "version": 1,
   "newProjectRoot": "projects",
   "projects": {
-    "interview_test_angular": {
+    "StudentApi": {
       "root": "",
       "sourceRoot": "src",
       "projectType": "application",
@@ -50,18 +50,18 @@
         "serve": {
           "builder": "@angular-devkit/build-angular:dev-server",
           "options": {
-            "browserTarget": "interview_test_angular:build"
+            "browserTarget": "StudentApi:build"
           },
           "configurations": {
             "production": {
-              "browserTarget": "interview_test_angular:build:production"
+              "browserTarget": "StudentApi:build:production"
             }
           }
         },
         "extract-i18n": {
           "builder": "@angular-devkit/build-angular:extract-i18n",
           "options": {
-            "browserTarget": "interview_test_angular:build"
+            "browserTarget": "StudentApi:build"
           }
         },
         "test": {
@@ -111,7 +111,7 @@
         }
       }
     },
-    "interview_test_angular-e2e": {
+    "StudentApi-e2e": {
       "root": "e2e/",
       "projectType": "application",
       "architect": {
@@ -119,7 +119,7 @@
           "builder": "@angular-devkit/build-angular:protractor",
           "options": {
             "protractorConfig": "e2e/protractor.conf.js",
-            "devServerTarget": "interview_test_angular:serve"
+            "devServerTarget": "StudentApi:serve"
           }
         },
         "lint": {
@@ -132,5 +132,5 @@
       }
     }
   },
-  "defaultProject": "interview_test_angular"
+  "defaultProject": "StudentApi"
 }
