+++ web/src/app/nav-menu/nav-menu.component.html
@@ -3,7 +3,7 @@
     class="navbar navbar-expand-sm navbar-toggleable-sm navbar-light bg-white border-bottom box-shadow mb-3"
   >
     <div class="container">
-      <a class="navbar-brand" [routerLink]="['/']">interview_test_angular</a>
+      <a class="navbar-brand" [routerLink]="['/']">StudentApi</a>
       <button
         class="navbar-toggler"
         type="button"
