+++ web/src/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <title>interview_test_angular</title>
+    <title>StudentApi</title>
     <base href="/" />
 
     <meta name="viewport" content="width=device-width, initial-scale=1" />
